+++ examples/basic/graphql-stream-transfers.ts
@@ -7,18 +7,16 @@ async function main() {
     network: DFUSE_API_NETWORK
   })
 
-  const stream = (await client.graphql<any>(streamTransfer, {
-    onMessage: (message) => {
-      if (message.type === "data") {
-        const actions = message.data.searchTransactionsForward.trace.matchingActions
+  const stream = await client.graphql(streamTransfer, (message) => {
+    if (message.type === "data") {
+      const actions = message.data.searchTransactionsForward.trace.matchingActions
 
-        actions.forEach(({ json }: any) => {
-          const { from, to, quantity, memo } = json
-          console.log(`Transfer [${from} -> ${to}, ${quantity}] (${memo})`)
-        })
-      }
+      actions.forEach(({ json }: any) => {
+        const { from, to, quantity, memo } = json
+        console.log(`Transfer [${from} -> ${to}, ${quantity}] (${memo})`)
+      })
     }
-  })) as Stream
+  })
 
   await waitFor(5000)
   await stream.close()